Total re-implementation with tests
